new folder structure + dependcy cleanup + import cleanup
