+++ vivid-server/src/models/channel.entity.ts
