+++ vivid-server/src/models/user.interface.ts
