+++ vivid-server/src/modules/app/app.module.ts
@@ -3,11 +3,11 @@ import { ConfigModule } from '@nestjs/config';
 import { ConfigService } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import configuration from './config/configuration';
+import configuration from '~/config/configuration';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { UserModule } from './users/user.module';
-import { ChannelModule } from './channels/channel.module';
+import { UserModule } from '$/users/user.module';
+import { ChannelModule } from '$/channels/channel.module';
 
 const config = ConfigModule.forRoot({
   load: [configuration],
