+++ vivid-server/src/modules/auth/auth.module.ts
@@ -1,7 +1,6 @@
 import { HttpModule, Module } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { PassportModule } from '@nestjs/passport';
-import { UserModule } from 'src/users/user.module';
+import { UserModule } from '$/users/user.module';
 import { AuthController } from './auth.controller';
 import { SessionSerializer } from './auth.serialize';
 import { AuthService } from './auth.service';
@@ -10,12 +9,7 @@ import { IntraStrategy } from './intra.strategy';
 @Module({
   controllers: [AuthController],
   providers: [IntraStrategy, SessionSerializer, AuthService, ConfigService],
-  imports: [
-    UserModule,
-    HttpModule,
-  ],
-  exports: [
-    AuthService
-  ]
+  imports: [UserModule, HttpModule],
+  exports: [AuthService],
 })
 export class AuthModule {}
