+++ vivid-server/src/modules/auth/auth.serialize.ts
@@ -1,12 +1,12 @@
 import { PassportSerializer } from '@nestjs/passport';
 import { Inject, Injectable } from '@nestjs/common';
-import { UserService } from 'src/users/user.service';
+import { UserService } from '$/users/user.service';
 
 @Injectable()
 export class SessionSerializer extends PassportSerializer {
   constructor(
     @Inject(UserService)
-    private readonly userService: UserService
+    private readonly userService: UserService,
   ) {
     super();
   }
@@ -15,7 +15,10 @@ export class SessionSerializer extends PassportSerializer {
     done(null, user.id.toString());
   }
 
-  async deserializeUser(id: string, done: (err: Error, user: any) => void): Promise<any> {
+  async deserializeUser(
+    id: string,
+    done: (err: Error, user: any) => void,
+  ): Promise<any> {
     try {
       const user = await this.userService.findUser(id);
       done(null, user);
