+++ vivid-server/src/modules/users/user.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserService } from './user.service';
-import { UserEntity } from './models/user.entity';
+import { UserEntity } from '@/user.entity';
 import { UserController } from './user.controller';
 
 @Module({
