+++ vivid-server/tsconfig.json
@@ -10,6 +10,11 @@
     "sourceMap": true,
     "outDir": "./dist",
     "baseUrl": "./",
+    "paths": {
+      "~/*": ["./src/*"],
+      "@/*": ["./src/models/*"],
+      "$/*": ["./src/modules/*"]
+    },
     "incremental": true
   }
 }
